# docs/api-spec.yml
...
openapi: 3.1.0
info:
  title: Archivo Desclasificado — Editorial Service
  version: 0.1.0
paths:
...
  /chapters/{chapter_id}/script/approve:
    post:
      tags:
      - script-approval
      summary: Approve Script
      description: 'Approve a chapter''s script, unblocking subtitle and video generation.


        Writes the decision to every platform version of the chapter, since they

        share one script. Returns 409 for a chapter with no platform versions —

        there would be nowhere to record the approval.'
      operationId: approve_script_chapters__chapter_id__script_approve_post
      parameters:
      - name: chapter_id
        in: path
        required: true
        schema:
          type: string
          title: Chapter Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScriptApprovalResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /chapters/{chapter_id}/script/reject:
    post:
      tags:
      - script-approval
      summary: Reject Script
      description: Withdraw approval, returning the chapter to the review queue.
      operationId: reject_script_chapters__chapter_id__script_reject_post
      parameters:
      - name: chapter_id
        in: path
        required: true
        schema:
          type: string
          title: Chapter Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScriptApprovalResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /chapters/{chapter_id}/script/update:
    post:
      tags:
      - script-approval
      summary: Update Script
      description: 'Replace the script text.


        Always resets approval, including when the chapter is already approved:

        the narrated text must be text a human read. Blank text is rejected with

        422, since an empty script would produce a silent video.'
      operationId: update_script_chapters__chapter_id__script_update_post
      parameters:
      - name: chapter_id
        in: path
        required: true
        schema:
          type: string
          title: Chapter Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScriptUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScriptApprovalResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /chapters/{chapter_id}/subtitles/generate:
    post:
      tags:
      - subtitles
      summary: Generate Subtitles
      description: 'Generate Spanish and English subtitle tracks.


        Timing is measured from the synthesized narration, not estimated. Returns

        409 unless the script is approved. Already-generated tracks are returned

        untouched, so calling twice cannot discard an editor''s corrections.'
      operationId: generate_subtitles_chapters__chapter_id__subtitles_generate_post
      parameters:
      - name: chapter_id
        in: path
        required: true
        schema:
          type: string
          title: Chapter Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubtitleTrackResponse'
                title: Response Generate Subtitles Chapters  Chapter Id  Subtitles
                  Generate Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /chapters/{chapter_id}/subtitles:
    get:
      tags:
      - subtitles
      summary: Get Subtitles
      description: Stored subtitle tracks; an empty list before generation.
      operationId: get_subtitles_chapters__chapter_id__subtitles_get
      parameters:
      - name: chapter_id
        in: path
        required: true
        schema:
          type: string
          title: Chapter Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubtitleTrackResponse'
                title: Response Get Subtitles Chapters  Chapter Id  Subtitles Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /chapters/{chapter_id}/subtitles/update:
    post:
      tags:
      - subtitles
      summary: Update Subtitles
      description: 'Replace segment text, preserving timing.


        Segments are matched by index and the count must equal the stored

        track''s, so timing cannot drift from the audio it was measured against.

        A mismatch or blank text is a 422.'
      operationId: update_subtitles_chapters__chapter_id__subtitles_update_post
      parameters:
      - name: chapter_id
        in: path
        required: true
        schema:
          type: string
          title: Chapter Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubtitleUpdateRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubtitleTrackResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ScriptApprovalResponse:
      properties:
        chapter_id:
          type: string
          title: Chapter Id
        script_approved:
          type: boolean
          title: Script Approved
        script_approved_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Script Approved At
        platform_versions_updated:
          type: integer
          title: Platform Versions Updated
      type: object
      required:
      - chapter_id
      - script_approved
      - script_approved_at
      - platform_versions_updated
      title: ScriptApprovalResponse
    ScriptUpdateRequest:
      properties:
        script_text:
          type: string
          minLength: 1
          title: Script Text
          description: Replacement script text
      type: object
      required:
      - script_text
      title: ScriptUpdateRequest
...
    SubtitleSegmentEditRequest:
      properties:
        index:
          type: integer
          title: Index
        text:
          type: string
          minLength: 1
          title: Text
      type: object
      required:
      - index
      - text
      title: SubtitleSegmentEditRequest
    SubtitleSegmentResponse:
      properties:
        index:
          type: integer
          title: Index
        start:
          type: string
          title: Start
        end:
          type: string
          title: End
        text:
          type: string
          title: Text
      type: object
      required:
      - index
      - start
      - end
      - text
      title: SubtitleSegmentResponse
    SubtitleTrackResponse:
      properties:
        lang:
          type: string
          title: Lang
        edited:
          type: boolean
          title: Edited
        updated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Updated At
        segments:
          items:
            $ref: '#/components/schemas/SubtitleSegmentResponse'
          type: array
          title: Segments
      type: object
      required:
      - lang
      - edited
      - updated_at
      - segments
      title: SubtitleTrackResponse
    SubtitleUpdateRequest:
      properties:
        lang:
          type: string
          title: Lang
        segments:
          items:
            $ref: '#/components/schemas/SubtitleSegmentEditRequest'
          type: array
          minItems: 1
          title: Segments
      type: object
      required:
      - lang
      - segments
      title: SubtitleUpdateRequest
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
servers:
- url: http://localhost:8000
  description: Local development server